feat: add the multiply function

+++ src/main/java/com/example/demo/arith/controller/ArithController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @RestController
 public class ArithController {
@@ -20,7 +21,7 @@ public class ArithController {
     @GetMapping("/add")
     public ResponseEntity<?> addTwoNumber (@PathVariable Long a, @PathVariable Long b) {
         try {
-            Long number = arithService.add(a,b);
+            BigInteger number = arithService.add(a,b);
             return ResponseEntity.ok().body(number);
         } catch (BadRequestException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
@@ -32,7 +33,7 @@ public class ArithController {
     @GetMapping("/minus")
     public ResponseEntity<?> minusTwoNumber (@PathVariable Long a, @PathVariable Long b) {
         try {
-            Long number = arithService.minus(a,b);
+            BigInteger number = arithService.minus(a,b);
             return ResponseEntity.ok().body(number);
         } catch (BadRequestException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
@@ -44,7 +45,7 @@ public class ArithController {
     @GetMapping("/multiply")
     public ResponseEntity<?> multiplyTwoNumber (@PathVariable Long a, @PathVariable Long b) {
         try {
-            Long number = arithService.multiply(a,b);
+            BigInteger number = arithService.multiply(a,b);
             return ResponseEntity.ok().body(number);
         } catch (BadRequestException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
